third_party/blink/web_tests/fast/xpath/4XPath/Borrowed/od_20000608-expected.txt

PASS nodeset.snapshotLength is 0
PASS nodeset.snapshotLength is 12
PASS successfullyParsed is true

TEST COMPLETE
